finishing refactoring Basic ATTs
