+++ src/NServiceBus.AcceptanceTests/Core/Conventions/When_sending_interface_message_with_conventions.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Basic
+﻿namespace NServiceBus.AcceptanceTests.Core.Conventions
 {
     using System;
     using System.Threading.Tasks;
