+++ src/NServiceBus.AcceptanceTests/Core/Pipeline/When_registering_handlers_explicitly.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Basic
+﻿namespace NServiceBus.AcceptanceTests.Core.Pipeline
 {
     using System.Threading.Tasks;
     using AcceptanceTesting;
