+++ src/NServiceBus.AcceptanceTests/Pipeline/When_handling_current_message_later.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Basic
+﻿namespace NServiceBus.AcceptanceTests.Pipeline
 {
     using System;
     using System.Threading.Tasks;
